[MIG]Commit inicial para migracion de POS a V19.
Se cambian como se comportan los modelos de python.
Reestructuracion de logica en componentes js,y modelos js.
Correcion de xpath en xml.
