+++ l10n_ve_pos/models/res_currency.py
@@ -7,10 +7,10 @@ class ResCurrency(models.Model):
     _inherit = 'res.currency'
 
     @api.model
-    def _load_pos_data_domain(self, data):
-        domain = super()._load_pos_data_domain(data)
-        company = self.env['res.company'].browse(data['pos.config']['data'][0]['company_id'])
-        currency_ids = [company.currency_id.id, data['pos.config']['data'][0]['currency_id']]
+    def _load_pos_data_domain(self, data, config):
+        domain = super()._load_pos_data_domain(data,config)
+        company = self.env['res.company'].browse(int(config['company_id']))
+        currency_ids = [company.currency_id.id, int(config['currency_id'])]
         if company.foreign_currency_id:
             currency_ids.append(company.foreign_currency_id.id)
         currency_ids = list(set(currency_ids))
