+++ l10n_ve_pos/static/src/overrides/components/closing_popup/closing_popup.xml
