+++ l10n_ve_pos/static/src/overrides/components/order_display/order_display.xml
@@ -1,5 +1,5 @@
 <templates id="template" xml:space="preserve">
-    <t t-name="l10n_ve_pos.OrderWidget" t-inherit="point_of_sale.OrderWidget" t-inherit-mode="extension">
+    <t t-name="l10n_ve_pos.OrderDisplay" t-inherit="point_of_sale.OrderDisplay" t-inherit-mode="extension">
         <!-- XPath para insertar los nuevos datos dentro del div 'order-summary' -->
         <xpath expr="//div[hasclass('order-summary')]" position="inside">
             <!-- Wrapper to align left and right content -->
