+++ l10n_ve_pos/static/src/overrides/components/orderline/orderline.xml
@@ -1,6 +1,6 @@
 <templates id="template" xml:space="preserve">
     <t t-name="l10n_ve_pos.Orderline" t-inherit="point_of_sale.Orderline" t-inherit-mode="extension">
-        <xpath expr="//div[hasclass('product-price')]" position="inside">
+        <xpath expr="//div[hasclass('product-price')]" position="after">
                 / <t t-out="line.foreignPrice"/>
         </xpath>
