+++ l10n_ve_pos/static/src/overrides/screens/product_screen/order_summary/order_sumarry.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates id="l10n_ve_pos_orderwidget_props" xml:space="preserve">
+  <t t-inherit="point_of_sale.OrderSummary" t-inherit-mode="extension">
+    <xpath expr="//OrderDisplay" position="attributes">
+            <attribute name="conversion_rate">currentOrder.get_conversion_rate()</attribute>
+            <attribute name="quantity_products">currentOrder.get_qty_products()</attribute>
+            <attribute name="foreign_total">env.utils.formatForeignCurrency(currentOrder.get_foreign_total_with_tax())</attribute>
+            <!-- 
+            <attribute name="foreign_tax">!env.utils.floatIsZero(currentOrder.get_foreign_total_tax()) and env.utils.formatForeignCurrency(currentOrder.get_foreign_total_tax()) or ''</attribute>
+            <attribute name="quantity_products">currentOrder.get_qty_products()</attribute>
+        <xpath expr="//ProductCard" position="attributes">
+            <attribute name="qty_available">product.qty_available</attribute>
+        </xpath> -->
+    </xpath>
+    </t>
+</templates>
