+++ l10n_ve_pos/static/src/overrides/screens/product_screen/product_list.js
@@ -1,0 +1,25 @@
+// /** @odoo-module **/
+
+// import { patch } from "@web/core/utils/patch";
+// import { ProductsWidget } from "@point_of_sale/app/screens/product_screen/product_list/product_list";
+
+// patch(ProductsWidget.prototype, {
+//   get productsToDisplay() {
+
+//     let list = super.productsToDisplay
+    
+//     // Filtrar productos si la configuración lo requiere
+//     if (!this.pos.config.pos_show_just_products_with_available_qty) {
+//       return list;
+//     }
+
+//     list = list.filter(product => {
+//       if (product.type === 'service' || product.type === 'consu') {
+//         return true;
+//       }
+//       return product.qty_available > 0;
+//     });
+
+//     return list;
+//   }
+// });
